+++ core-ng/src/main/java/core/framework/http/HTTPClientException.java
@@ -2,10 +2,13 @@ package core.framework.http;
 
 import core.framework.log.ErrorCode;
 
+import java.io.Serial;
+
 /**
  * @author neo
  */
 public final class HTTPClientException extends RuntimeException implements ErrorCode {
+    @Serial
     private static final long serialVersionUID = -7577571026619549539L;
 
     private final String errorCode;