+++ core-ng/src/main/java/core/framework/module/DBConfig.java
@@ -4,6 +4,7 @@ import core.framework.db.Database;
 import core.framework.db.IsolationLevel;
 import core.framework.db.Repository;
 import core.framework.internal.db.DatabaseImpl;
+import core.framework.internal.db.cloud.GCloudAuthProvider;
 import core.framework.internal.module.Config;
 import core.framework.internal.module.ModuleContext;
 import core.framework.internal.module.ShutdownHook;
@@ -58,7 +59,11 @@ public class DBConfig extends Config {
     }
 
     public void user(String user) {
-        database.user = user;
+        if ("gcloud:iam".equals(user)) {
+            database.authProvider = new GCloudAuthProvider();
+        } else {
+            database.user = user;
+        }
     }
 
     public void password(String password) {